[fix]: do not use fixed ports in tests + use multi-threading rather that multi-processing to spawn http tests
