+++ wayflowcore/tests/tracing/opentelemetry/conftest.py
@@ -3,17 +3,83 @@
 # This software is under the Apache License 2.0
 # (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0) or Universal Permissive License
 # (UPL) 1.0 (LICENSE-UPL or https://oss.oracle.com/licenses/upl), at your option.
+import json
+import threading
+from http.server import BaseHTTPRequestHandler, HTTPServer
+from typing import Any, Dict
 
 import pytest
 
-from .otel_server_utils import _start_otel_server
+from ...utils import get_available_port
 
-OTEL_SERVER_PORT = 4318
+
+class SimpleOTLPHandler(BaseHTTPRequestHandler):
+
+    stored_spans: Dict[int, Any] = {}
+
+    def _traces(self) -> None:
+        response_status = 200
+        content_length = int(self.headers.get("Content-Length", 0))
+        body = self.rfile.read(content_length)
+
+        try:
+            trace_data = json.loads(body)
+            span_id = int(trace_data["context"]["span_id"], 16) - 1
+            SimpleOTLPHandler.stored_spans[span_id] = trace_data
+        except Exception:
+            response_status = 400
+
+        self.send_response(response_status)
+        self.end_headers()
+        self.wfile.write(b"")  # Empty response
+
+    def _get_span(self):
+        response_status = 200
+        content_length = int(self.headers.get("Content-Length", 0))
+        body = self.rfile.read(content_length)
+
+        try:
+            request_data = json.loads(body)
+            trace_data = json.dumps(SimpleOTLPHandler.stored_spans[int(request_data["span_id"])])
+        except Exception:
+            trace_data = "Provided body data was in wrong format"
+            response_status = 400
+
+        self.send_response(response_status)
+        self.end_headers()
+        self.wfile.write(trace_data.encode())
+
+    def do_GET(self) -> None:
+        self.send_response(404)
+        self.end_headers()
+
+    def do_POST(self) -> None:
+        if self.path == "/v1/traces":
+            self._traces()
+        elif self.path == "/v1/getspan":
+            self._get_span()
+        else:
+            self.send_response(404)
+            self.end_headers()
 
 
 @pytest.fixture(scope="session")
 def otel_server():
-    process, url = _start_otel_server(host="localhost", port=OTEL_SERVER_PORT)
-    yield url
-    process.kill()
-    process.wait()
+    host = "localhost"
+    port = get_available_port()
+    url = f"{host}:{port}"
+
+    server = HTTPServer((host, port), SimpleOTLPHandler)
+
+    thread = threading.Thread(
+        target=server.serve_forever,
+        name="otel-test-server",
+        daemon=True,
+    )
+    thread.start()
+    try:
+        yield url
+    finally:
+        server.shutdown()
+        server.server_close()
+        thread.join(timeout=5)
